backend/requirements.txt
fastapi
pydantic
python-jose[cryptography]
uvicorn
python-multipart
psycopg2-binary
passlib[bcrypt]
python-dotenv
bcrypt